Return boolean in deleteById instead of void

+++ module2web/src/main/java/com/spring/abir/module2web/services/EmployeeService.java
@@ -46,7 +46,10 @@ public class EmployeeService {
         return modelMapper.map(savedEmployeeEntity, EmployeeDTO.class);
     }
 
-    public void deleteEmployeeById(Long id) {
+    public boolean deleteEmployeeById(Long id) {
+        boolean exists = employeeRepository.existsById(id);
+        if(!exists) return false;
         employeeRepository.deleteById(id);
+        return true;
     }
 }
